+++ CHANGES.md
@@ -4,9 +4,12 @@ Released: -
 
 - Remove the validation of input locations ([issue #259][issue_259]).
 - Support passing `Function` type config to `SWAGGER_UI_CONFIG` ([issue #381][issue_381]).
+- Fix the base response support so that a custom class can be returned from
+  the view function ([issue #384][issue_384]).
 
 [issue_259]: https://github.com/apiflask/apiflask/issues/259
 [issue_381]: https://github.com/apiflask/apiflask/issues/381
+[issue_384]: https://github.com/apiflask/apiflask/issues/384
 
 
 ## Version 1.2.1